working, but still not saving value
