+++ core/src/main/java/no/nordicsemi/android/common/core/settings/NordicCommonLibsSettingsRepository.kt
@@ -2,19 +2,38 @@ package no.nordicsemi.android.common.core.settings
 
 import android.content.Context
 import android.util.Log
+import androidx.datastore.core.DataStore
+import androidx.datastore.dataStore
+import dagger.hilt.android.qualifiers.ApplicationContext
 import kotlinx.coroutines.flow.Flow
 import kotlinx.coroutines.flow.catch
 import kotlinx.coroutines.flow.first
+import no.nordicsemi.android.common.core.settings.migrations.getMigrationsList
 import java.io.IOException
+import javax.inject.Inject
+import javax.inject.Singleton
 
-class SettingsRepository(private val context: Context) {
-    private val dataStore = context.nordicSettingsDataStore
-    val nordicSettings: Flow<Settings> = dataStore.data
+private const val DATASTORE_FILENAME = "nordic_common_libs_settings.pb"
+val Context.nordicCommonLibsSettingsDataStore: DataStore<NordicCommonLibsSettings> by
+dataStore (
+    fileName = DATASTORE_FILENAME,
+    serializer = NordicCommonLibsSettingsSerializer,
+    produceMigrations = { context -> getMigrationsList(context) }
+)
+
+@Singleton
+class NordicCommonLibsSettingsRepository @Inject constructor(
+    @ApplicationContext context: Context
+) {
+    private val dataStore by lazy {
+        context.nordicCommonLibsSettingsDataStore
+    }
+    val nordicCommonLibsSettings: Flow<NordicCommonLibsSettings> = dataStore.data
         .catch { exception ->
             // dataStore.data throws an IOException when an error is encountered when reading data
             if (exception is IOException) {
                 Log.e(TAG, "Error reading $DATASTORE_FILENAME.", exception)
-                emit(Settings.getDefaultInstance())
+                emit(NordicCommonLibsSettings.getDefaultInstance())
             } else {
                 throw exception
             }
@@ -29,7 +48,7 @@ class SettingsRepository(private val context: Context) {
         }
     }
 
-    suspend fun updateBluetoothPermissionsRequested(requested: Boolean) {
+    suspend fun updateBluetoothPermissionRequested(requested: Boolean) {
         dataStore.updateData { settings ->
             settings
                 .toBuilder()
