+++ core/src/main/proto/NordicCommonLibSettings.proto
@@ -3,7 +3,7 @@ syntax = "proto3";
 option java_package = "no.nordicsemi.android.common.core.settings";
 option java_multiple_files = true;
 
-message Settings {
+message NordicCommonLibsSettings {
 
   bool location_permission_requested = 1;
   bool bluetooth_permission_requested = 2;
