+++ permissions-ble/src/main/java/no/nordicsemi/android/common/permissions/ble/RequireBluetooth.kt
@@ -39,7 +39,7 @@ import androidx.compose.runtime.LaunchedEffect
 import androidx.compose.runtime.getValue
 import androidx.hilt.navigation.compose.hiltViewModel
 import androidx.lifecycle.compose.collectAsStateWithLifecycle
-import no.nordicsemi.android.common.permissions.ble.util.BlePermissionState
+import no.nordicsemi.android.common.permissions.ble.permissionState.BlePermissionState
 import no.nordicsemi.android.common.permissions.ble.view.BluetoothDisabledView
 import no.nordicsemi.android.common.permissions.ble.view.BluetoothNotAvailableView
 import no.nordicsemi.android.common.permissions.ble.view.BluetoothPermissionRequiredView
