+++ permissions-ble/src/main/java/no/nordicsemi/android/common/permissions/ble/viewmodel/PermissionViewModel.kt
@@ -38,9 +38,9 @@ import dagger.hilt.android.lifecycle.HiltViewModel
 import kotlinx.coroutines.flow.SharingStarted
 import kotlinx.coroutines.flow.stateIn
 import no.nordicsemi.android.common.permissions.ble.BlePermissionNotAvailableReason
-import no.nordicsemi.android.common.permissions.ble.bluetooth.BluetoothStateManager
-import no.nordicsemi.android.common.permissions.ble.location.LocationStateManager
-import no.nordicsemi.android.common.permissions.ble.util.BlePermissionState
+import no.nordicsemi.android.common.permissions.ble.permissionState.BluetoothStateManager
+import no.nordicsemi.android.common.permissions.ble.permissionState.LocationStateManager
+import no.nordicsemi.android.common.permissions.ble.permissionState.BlePermissionState
 import javax.inject.Inject
 
 /**
