+++ permissions-notification/build.gradle.kts
@@ -29,6 +29,7 @@ dokka {
 dependencies {
     implementation(project(":core"))
     implementation(libs.accompanist.permissions)
+    implementation(libs.androidx.datastore)
     implementation(libs.protobuf.javalite)
 
     implementation(libs.androidx.compose.material.iconsExtended)
