+++ permissions-wifi/build.gradle.kts
@@ -60,6 +60,7 @@ dokka {
 dependencies {
     implementation(project(":core"))
     implementation(project(":ui"))
+    implementation(libs.androidx.datastore)
     implementation(libs.protobuf.javalite)
 
     implementation(libs.androidx.compose.material.iconsExtended)
