+++ permissions-wifi/src/main/java/no/nordicsemi/android/common/permissions/wifi/RequireWiFi.kt
@@ -39,7 +39,7 @@ import androidx.compose.runtime.LaunchedEffect
 import androidx.compose.runtime.getValue
 import androidx.hilt.navigation.compose.hiltViewModel
 import androidx.lifecycle.compose.collectAsStateWithLifecycle
-import no.nordicsemi.android.common.permissions.wifi.utils.WiFiPermissionState
+import no.nordicsemi.android.common.permissions.wifi.permissionState.WiFiPermissionState
 import no.nordicsemi.android.common.permissions.wifi.view.WiFiDisabledView
 import no.nordicsemi.android.common.permissions.wifi.view.WiFiNotAvailableView
 import no.nordicsemi.android.common.permissions.wifi.view.WiFiPermissionRequiredView
